slack/requirements.txt
SQLAlchemy<2.0.0
psycopg2-binary
requests